refactor(popup): demo调整
